+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `TemplateEngine` enum
 - `MemorError` class
 - `PromptTemplate` class `render` method
 - `PromptTemplate` class `from_content_file` method
+- `PromptTemplate` class `update_engine` method
+- `PromptTemplate` class `engine` property
 ### Changed
+- `engine` parameter added to `PromptTemplate` class
 - `Prompt` class `render` method updated
 - `LLMModel` structure updated
 - `LLMModel` type changed